+++ src/app/DTT/text/index.js
@@ -1,23 +1,23 @@
-import { useEffect, useRef, useState } from 'react'
+import { ACKMESSAGEKEY, maxNumberOfRetries } from 'utils/constants'
 import { Box, Text as ChakraText, HStack, Input, VStack, useTheme } from '@chakra-ui/react'
-import { useSelector } from 'react-redux'
 import { faCalendar, faCheckCircle } from '@fortawesome/free-solid-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { useEffect, useRef, useState } from 'react'
 
 import { ACTIONS } from 'utils/contexts/ErrorReducer'
+import AnswerAcknowledge from 'app/layouts/components/form/answer_acknowledge'
 import DetailViewTags from 'app/DTT/text/detailview_tags'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import InputMask from 'react-input-mask'
+import MandatorySymbol from 'app/layouts/components/form/mandatory-symbol'
 import debounce from 'lodash.debounce'
 import { getIsInvalid } from 'utils/functions'
 import { isNotNullOrUndefinedOrEmpty } from 'utils/helpers/is-null-or-undefined'
+import { selectCode } from 'redux/db/selectors'
 import { useError } from 'utils/contexts/ErrorContext'
 import useGetFieldMessage from 'utils/fieldMessage'
 import { useIsFieldNotEmpty } from 'utils/contexts/IsFieldNotEmptyContext'
 import useProductColors from 'utils/productColors'
-import { selectCode } from 'redux/db/selectors'
-import { maxNumberOfRetries, ACKMESSAGEKEY } from 'utils/constants'
-import AnswerAcknowledge from 'app/layouts/components/form/answer_acknowledge'
-import MandatorySymbol from 'app/layouts/components/form/mandatory-symbol'
-import InputMask from 'react-input-mask'
+import { useSelector } from 'react-redux'
 
 export const Write = ({
   questionCode,
